Fix: Align MapMessage protocol for map rendering
The client was failing to render the map correctly due to a data type mismatch in the MapMessage protocol.

The server was sending map coordinates (x, y) and dimensions (width, height) as 4-byte integers, while the reference 'duskz' protocol uses 2-byte shorts.

This change updates the MapMessage class on both the client and server to use 'short' for these fields, ensuring consistent serialization and deserialization and fixing the rendering bug.

+++ DuskFiles/ClientSources/Clientz2/MapMessage.java
@@ -3,10 +3,10 @@ import java.io.DataOutputStream;
 import java.io.IOException;
 
 public class MapMessage extends DuskMessage {
-	public int x;
-	public int y;
-	public int width;
-	public int height;
+	public short x;
+	public short y;
+	public short width;
+	public short height;
 	public short[][][] map;
 
 	public MapMessage() {
@@ -14,18 +14,18 @@ public class MapMessage extends DuskMessage {
 
 	public MapMessage(byte name, int x, int y, int width, int height) {
 		super(name);
-		this.x = x;
-		this.y = y;
-		this.width = width;
-		this.height = height;
+		this.x = (short)x;
+		this.y = (short)y;
+		this.width = (short)width;
+		this.height = (short)height;
 		this.map = new short[3][width][height];
 	}
 
 	@Override
 	public void receive(DataInputStream istream) throws IOException {
 		super.receive(istream);
-		x = istream.readInt();
-		y = istream.readInt();
+		x = istream.readShort();
+		y = istream.readShort();
 		width = istream.readShort();
 		height = istream.readShort();
 		map = new short[3][width][height];
@@ -41,8 +41,8 @@ public class MapMessage extends DuskMessage {
 	@Override
 	public void send(DataOutputStream out) throws IOException {
 		super.send(out);
-		out.writeInt(x);
-		out.writeInt(y);
+		out.writeShort(x);
+		out.writeShort(y);
 		out.writeShort(width);
 		out.writeShort(height);
 		for (int l = 0; l < 3; l++) {
